requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
alembic
python-dotenv
pandas
openpyxl
transformers>=4.30.0
torch>=2.0.0
spacy>=3.5.0
underthesea>=1.3.0
simalign>=0.3.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0 